Refactor: 그룹 장바구니 헤더의 ProgressBar 컴포넌트 분리 및 애니메이션 추가
- ProgressBar에 애니메이션 효과를 추가하여 진행률 변경이 부드럽게 표시되도록 개선했습니다.
- ProgressBar의 모서리를 둥글게 처리하여 디자인을 개선했습니다.
