+++ feature/group-cart/src/main/java/com/cyberwarriers/zubzub/feature/group_cart/presentation/ui/components/GroupCartHeader.kt
@@ -65,15 +65,4 @@ fun GroupCartHeaderPreview() {
             progress = 0.6f
         )
     }
-}
-
-@Preview
-@Composable
-fun ProgressBarPreview() {
-    ZubZubTheme {
-        ProgressBar(
-            progress = 0.75f,
-            modifier = Modifier.padding(16.dp)
-        )
-    }
 }
